reroute client routes to clients file for organization
